Move test Model to separate files
