+++ Tests/ContentfulTests/QueryTests.swift
@@ -10,87 +10,6 @@
 import XCTest
 import DVR
 
-final class Cat: Resource, EntryDecodable, FieldKeysQueryable {
-
-    static let contentTypeId: String = "cat"
-
-    let sys: Sys
-    let color: String?
-    let name: String?
-    let lives: Int?
-    let likes: [String]?
-
-    // Relationship fields.
-    var bestFriend: Cat?
-    var image: Asset?
-
-    public required init(from decoder: Decoder) throws {
-        sys             = try decoder.sys()
-        let fields      = try decoder.contentfulFieldsContainer(keyedBy: Cat.FieldKeys.self)
-
-        self.name       = try fields.decodeIfPresent(String.self, forKey: .name)
-        self.color      = try fields.decodeIfPresent(String.self, forKey: .color)
-        self.likes      = try fields.decodeIfPresent(Array<String>.self, forKey: .likes)
-        self.lives      = try fields.decodeIfPresent(Int.self, forKey: .lives)
-
-        try fields.resolveLink(forKey: .bestFriend, decoder: decoder) { [weak self] linkedCat in
-            self?.bestFriend = linkedCat as? Cat
-        }
-        try fields.resolveLink(forKey: .image, decoder: decoder) { [weak self ] image in
-            self?.image = image as? Asset
-        }
-    }
-    
-    enum FieldKeys: String, CodingKey {
-        case bestFriend, image
-        case name, color, likes, lives
-    }
-}
-
-final class City: Resource, EntryDecodable, FieldKeysQueryable {
-
-    static let contentTypeId: String = "1t9IbcfdCk6m04uISSsaIK"
-
-    let sys: Sys
-    var location: Location?
-
-    public required init(from decoder: Decoder) throws {
-        sys             = try decoder.sys()
-        let fields      = try decoder.contentfulFieldsContainer(keyedBy: City.FieldKeys.self)
-
-        self.location   = try fields.decode(Location.self, forKey: .location)
-    }
-
-    enum FieldKeys: String, CodingKey {
-        case location = "center"
-    }
-}
-
-final class Dog: Resource, EntryDecodable, FieldKeysQueryable {
-
-    static let contentTypeId: String = "dog"
-
-    let sys: Sys
-    let name: String!
-    let description: String?
-    var image: Asset?
-
-    public required init(from decoder: Decoder) throws {
-        sys             = try decoder.sys()
-        let fields      = try decoder.contentfulFieldsContainer(keyedBy: Dog.FieldKeys.self)
-        name            = try fields.decode(String.self, forKey: .name)
-        description     = try fields.decodeIfPresent(String.self, forKey: .description)
-
-        try fields.resolveLink(forKey: .image, decoder: decoder) { [weak self] linkedImage in
-            self?.image = linkedImage as? Asset
-        }
-    }
-
-    enum FieldKeys: String, CodingKey {
-        case image, name, description
-    }
-}
-
 class QueryTests: XCTestCase {
 
     static let client: Client = {
